requirements.txt
torch
torchaudio
openai-whisper
ffmpeg
